check in call_run if fields are of compatible type and layout, which requires moving the StencilObject class from definitions.py to a new separate file (stencil_object.py)
