+++ src/gt4py/backend/debug_backend.py
@@ -14,6 +14,8 @@
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 
+import numpy as np
+
 from gt4py import backend as gt_backend
 from gt4py import ir as gt_ir
 from gt4py import definitions as gt_definitions
@@ -207,6 +209,10 @@ def debug_is_compatible_layout(field):
     return sum(field.shape) > 0
 
 
+def debug_is_compatible_type(field):
+    return isinstance(field, np.ndarray)
+
+
 @gt_backend.register
 class DebugBackend(gt_backend.BaseBackend):
     name = "debug"
@@ -216,6 +222,7 @@ class DebugBackend(gt_backend.BaseBackend):
         "device": "cpu",
         "layout_map": debug_layout,
         "is_compatible_layout": debug_is_compatible_layout,
+        "is_compatible_type": debug_is_compatible_type,
     }
 
     GENERATOR_CLASS = DebugGenerator
