+++ src/gt4py/backend/numpy_backend.py
@@ -14,6 +14,8 @@
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 
+import numpy as np
+
 from gt4py import backend as gt_backend
 from gt4py import ir as gt_ir
 from gt4py import definitions as gt_definitions
@@ -296,6 +298,10 @@ def numpy_is_compatible_layout(field):
     return sum(field.shape) > 0
 
 
+def numpy_is_compatible_type(field):
+    return isinstance(field, np.ndarray)
+
+
 @gt_backend.register
 class NumPyBackend(gt_backend.BaseBackend):
     name = "numpy"
@@ -305,6 +311,7 @@ class NumPyBackend(gt_backend.BaseBackend):
         "device": "cpu",
         "layout_map": numpy_layout,
         "is_compatible_layout": numpy_is_compatible_layout,
+        "is_compatible_type": numpy_is_compatible_type,
     }
 
     GENERATOR_CLASS = NumPyGenerator
